Added some tests, moved GraphUtils to own file
